\name{RChurch-package}
\alias{RChurch-package}
\alias{RChurch}
\docType{package}
\title{
R interface to Church
}
\description{
Provides an interface to the Church set of Scheme commands. Allows models to be specified as R scripts and 
translates the scripts to Church behind the scenes.
}
\details{
\tabular{ll}{
Package: \tab RChurch\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-08-11\cr
License: \tab GPL 3\cr
LazyLoad: \tab yes\cr
}
See \code{\link{church.model}} and \code{\link{church.sample}}. Run \code{vignette{'RChurch'}} for an overview and some examples.
}
\author{
Jon Malmaud <malmaud@mit.edu>
}
